+++ modules/documentation/docs/changelog.md
@@ -3,6 +3,7 @@
 ## Next Release
 
 - fix browser node store issue
+- fix `TFullTransferState` schema
 
 ## 0.0.28